pyproject.toml
[tool.poetry]
name = "fake-monet"
version = "0.1.0"
description = ""
authors = ["Gabriel Leite Bessa <gabrielleitebessa@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"
tensorflow = "^2.17.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
